Ana Tasarım Hatları Yapıldı.Firebase Auth, Navigation gibi temel düzenlemeler ve eklemeler yapıldı... Kullanıcı bilgileri, döngüler ve mevcut kişi gibi bilgiler textViewlara aktarıldı. İşlemler Tamamlandı.

+++ app/src/main/java/com/example/bankproject/SiraKayitFragment.kt
@@ -5,9 +5,11 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ArrayAdapter
 import androidx.navigation.Navigation
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.FirebaseDatabase
+import kotlinx.android.synthetic.main.fragment_sira_kayit.*
 import kotlinx.android.synthetic.main.fragment_sira_kayit.view.*
 
 
@@ -21,6 +23,7 @@ class SiraKayitFragment : Fragment() {
         database = FirebaseDatabase.getInstance()
 
 
+
         tasarim.exitButtonYonetici.setOnClickListener {
             auth.signOut()
             Navigation.findNavController(it).navigate(R.id.action_siraKayitFragment_to_yoneticiGirisFragment)
@@ -28,10 +31,10 @@ class SiraKayitFragment : Fragment() {
 
 
         tasarim.buttonYoneticiKaydet.setOnClickListener {
-            val ad_soyad = tasarim.yoneticiKullaniciAd.text.toString().trim()
+          val ad_soyad = tasarim.yoneticiKullaniciAd.text.toString().trim()
             val gmail = tasarim.yoneticiMusteriGmail.text.toString().trim()
             val tc_no = tasarim.yoneticiMusteriKimlikNumarasi.text.toString().trim().toLong()
-            val gise_no = tasarim.yoneticiMusteriGiseNo.text.toString().trim().toInt()
+            val gise_no = tasarim.yoneticiMusteriGiseNo.text.toString().trim()
             val sira_no = tasarim.yoneticiMusteriSiraNo.text.toString().trim().toInt()
 
             val refMusteriler = database.getReference("musteriler")
@@ -46,4 +49,5 @@ class SiraKayitFragment : Fragment() {
         return tasarim
     }
 
-}
+
+    }
